Fix concurrency in comment report calls
The comments reporting mechanism was implemented from a listener
registered in the list adapter implemented by the
CourseDiscussionCommentsFragment instance itself in commit
7821e17e4665c1e8d20f18978eefbe89f688289c. The implementation attempted
to block concurrent calls for reporting a comment, that could caused by
the user clicking multiple times on the button, by cancelling the
previous ongoing call whenever it was invoked. However, since only one
call was stored without any mapping to the actual comment, this caused
a report action on one comment to interrupt the ongoing action on
another comment instead.

For now this commit addresses this by just removing the logic to cancel
the previous ongoing call. This should mostly work in accordance with
the user’s expectations. The proper way to do this would be to have the
report button disabled on any comment where the report call is ongoing,
but this issue needs to be addressed in the other modules (threads,
responses) as well.