Fix Retrofit call cancellation triggering error
OkHttp throws a standard IOException with the message "Canceled"
whenever any call is cancelled, which is passed on by the base
ErrorHandlingCallback implementation to the ErrorUtils.getErrorMessage()
method, which in turn interprets it as a network connectivity issue, and
returns the appropriate error message. This commit fixes the issue by
adding a check in the ErrorHandlingCallback implementation’s failure
callback to verify that the call hasn’t been cancelled, before
delivering the error message.