Remove courseware menu
The courseware module was implementing some custom popup menus on the
action bar, for viewing a course on the web and toggling a filter to
show only video courses. These menus were implemented using a custom
popup menu implementation with non-standard modifications.

These menus are now mostly obsolete. We want to encourage users to view
the courses in the app itself, and any course that isn’t supported on
the app would already have a full-screen message to that effect with a
button that redirects to the browser. The video filtering menu was also
used very little overall.

This commit removes these menus along with their functionality. The
custom menu framework that was implemented for these menus was also
being used with the course sharing and photo capture buttons. This usage
was incorrect, because the custom menu was customized specifically for
the action bar menu. Therefore the menu implementation is being changed
to the default one from the appcompat library in the other cases.
(Technically, these button popups in the content view should be using
Spinner instead of PopupMenu.)