Add Bulk Videos Download feature
- LEARNER-2940
- Add ConstraintLayout and use it for building the bulk videos download
view
Database is queried for all the data that it has for a course's videos
and then they are cross matched with the videos data of the same course
that we get from the server call to decide:
1 - How many videos have been downloaded
2 - How many videos can be downloaded
3 - How many videos are currently being downloaded

Two lists are kept for the implementation to work.
- 1st list keeps all the remaining videos that can be downloaded.
- 2nd list keeps all the videos that have been downloaded or the ones
that are currently being downloaded.
When the switch is ON, all videos in the 1st list are downloaded.
When the switch is OFF, all videos in the 2nd list are deleted (if
downloaded) and stopped (if downloading) after a delay.

When all videos are in progress, tapping the bulk download view takes
user to the Download Videos screen.

- Helpers for getting progress of downloading videos in a course
Functions added in various classes to obtain the videos that are
being downloaded by the DownloadManager, filtering them on the
basis of Course ID.
If Course ID isn't provided, unfiltered videos being downloaded
are returned.
- Only show download progress menu item on Course Tab
- Hide bulk download view if a course has NO downloadable videos
- Add accessibility to bulk download view
- Put all business logic in AsyncTask for DownloadCompleteReceiver
The code inside receiver was previously being run on the main thread
and didn't involve any interaction with the UI elements causing lag
within the app.

Introduce fragment for Bulk Download View

- BulkDownloadFragment is responsible for handling all the various
states of the bulk download view.
- Switch's state is saved in the prefs based on the course that its
being set in.
- Whenever a video is deleted, the bulk download switch's state for the
course whose video is deleted gets reset in the prefs.
- Disable BulkDownload while videos are being queued by DownloadManager
to download
- Update the download state of ListView rows whenever a download is
initiated even on a different tab
- Hide progress on BulkDownload view whenever a video is deleted or
an ongoing video download is stopped
- Spinner on Downloads screen before ListView is populated
- Update switch's state when it is OFF and all videos are individually
enqueued for downloading
- Commulative progress instead of the one download manager gives us
- Standardize the download limit check to DownloadUtil class's
isDownloadSizeWithinLimit function

Accessibility Improvements
- "Bulk videos download switch" added as a prefix before adding title,
subtitle etc to switch's contentDescription.
- While the switch is in ON state the whole view is now tappable to
during accessibility to open the Downloads screen. Matching string as
iOS in this case has also been added.