Download and Delete implementation for Switch
Two lists are kept for the implementation to work.
- 1st list keeps all the remaining videos that can be downloaded.
- 2nd list keeps all the videos that have been downloaded or the ones
that are currently being downloaded.
When the switch is ON, all videos in the 1st list are downloaded.
When the switch is OFF, all videos in the 2nd list are deleted (if
downloaded) and stopped (if downloading) after a delay.

When all videos are in progress, tapping the bulk download view takes
user to the Download Videos screen.