Handle the case where server sends non-parsable course content
- LEARNER-3785
Previously our app used to crash whenever the server returned a course
whose content was invalid and logged an exception on Crashlytics. This
commit gracefully handles that exception by wrapping it into
CourseContentNotValidException and showing an error on screen while
sending the exception to Crashlytics in the background too.

CourseContentNotValidException is either thrown when raised during a
test's execution or caught and re-thrown again as a RuntimeException as
it was being done for some other cases as well.