Add initial support of landscape mode in some screens.
> LEARNER-3384
> Enable landscape mode in MainDashboardActivity, CourseTabsDashboardActivity,
CourseOutlineActivity, CourseHandoutActivity, CourseAnnouncementActivity and
DownloadListActivity screens.
> Done following fixes after enabling landscape mode on these screens
- Stretch tabs of tabs layout to max of its available width.
- Don’t show ‘Find a course’ card on ‘Courses’ fragment until the screen loads.
- Fixed spiral progress bar stops redrawing itself issue. The issue occurs in
spiral progress bar for video downloading, it works fine when in one orientation
(either portrait or landscape) but when you change the orientation it stops
redrawing itself.
- Update progress wheel library, update is avoiding use of handler.
- Avoid the recreation of DownloadSizeExceed Dialog on orientation change.
As a quick fix, avoid the recreation of dialog on orientation change to
avoid problematic scenarios For e.g. cancel button stops working on
orientation change, we have to fix it properly which is explained and
planned to implement in story LEARNER-2177.
- Removing all downloads from native download manager initially.
It has been done to avoid possible issues for example: #1106 (comment)