Fix TransactionTooLargeException on unit navigation
- LEARNER-6680
TransactionTooLargeException happens when an Activity/Fragment is in the
process of stopping, that means that the Activity/Fragment is trying to
send its saved state Bundles to the system OS for safe keeping for
restoration later (after a config change or process death) but that one
or more of the Bundles it sent is too large (i.e. greater than 1MB).
To avoid this scenario on CourseUnitNavigationActivity, CourseComponent
object is now saved on the disk using the Bridge library which allows
writing and retrieving it from the disk.