Add empty google-services.json file
For our project to properly compile we need to have a `google-services.json` file because we are using Firebase in the project which requires this file. By having an empty `google-services.json` file, we are allowing the community to use the project even when they don't wish to use Firebase on their ends.