Stop cache clearing for all WebViews
Calling the clearCache function on any WebView within the app clears the
cache through out the app which isn't what we want.
Our default strategy for caching is LOAD_DEFAULT since nothing has been
explicitly set in our EdxWebView class, which ensures that for a normal
page load, the cache is checked and content is revalidated if needed.