Fix WebView crash related to hardware acceleration
- LEARNER-7295
The problem is caused when a WebView is placed inside a ScrollView or
NestedScrollView while hardware acceleration is enabled.
This is likely a bug with how ScrollView calculates the width and
height of a WebView inside it and assigns a very large number.
Arbitrarily large views can't be put into hardware layers. Hardware
layers are backed by textures, and GPU drivers have a predefined maximum
size for textures.
For more info:
https://bugs.chromium.org/p/chromium/issues/detail?id=578150