Handle a case where activity is destroyed when user leaves receipt page
Workaround to handle the case where a user has successfully made the
payment and has reached the receipt page and somehow they leave the
screen and OS decides to destroy the activity.
If this workaround is not in place, upon return to the app, the user
will again see the payment page presenting them with a form for the
payment to upgrade a course.