Configure EULA, TOS & Privacy Policy independently
- LEARNER-7763
Links to these agreements can now be provided through configurations.
When available, these URLs will always have precedance over the local
files we already have in our project.
A catch here is that if we only provide one URL through configurations
then only that agreement's text will be clickable and the others won't
be clickable even if the local files are available in the project.