Remove chevrons from clickable list items
- LEARNER-8254

While this isn’t something that’s explicitly called out anywhere in
Android or Material guidelines but it's kind of a de-facto iOS-only
thing to have a chevron to show clickable row in a list.

Here’s a relevant article that explains this:
https://medium.com/@mobiledesignguy/chevrons-in-native-android-apps-36955564c938

p.s. Material Guidelines don’t include a chevron in any of their
sample images here:
https://material.io/archive/guidelines/components/lists.html