feat: add deep-link to the component that is due in my calendar event
Add deep-link to the calendar event and navigate to the course
component that is due.

Fix issue of showing calendar outdated modal. After adding the
deep-link in the calendar event, the App shows event outdated
model cuz app matching the event description and the time to
check that events are the updated from the outside of the app
or not.

fixes: LEARNER-8298

issue fixed by use of `contains` string method instated
of the `equals`.