fix: Error dialog not appearing correctly on Course Component screen
The component screen was unable to observe the event because we
consumed the shared event on CourseUnitNavigationActivity. This
resulted in the event being unavailable for the screen to receive.
To prevent a recurrence of this issue, we have taken steps to ensure
that events are only consumed if they are specifically related to
that screen.

Fixes: LEARNER-9314