fix: TransactionTooLarge Exception caused by FragmentStateAdapter
Previously, the FragmentStateAdapter was saving the state of the
course dashboard view pager tabs upon navigation to the
NavigationActivity. This behavior resulted in a TransactionTooLarge
Exception when attempting to save the state in the bundle.

To mitigate this issue, we've made adjustments to the caching
mechanism of the ViewPager2. Specifically, we've set the OffPage
limit of the viewPager to ViewPager2.OFFSCREEN_PAGE_LIMIT_DEFAULT,
which leverages the default caching mechanism of RecyclerView.
This modification ensures efficient state handling and resolves the
encountered exception.

Fixes: LEARNER-9897